Docs: Adjust markdown around shared content list
The organization preferences page is unexpectedly rendering an `<ol>`
tag in a codeblock where the `docs/list` shortcode is in use. Adjust
the Markdown around these lists to remove line breaks and indents
around those shortcodes to attempt to remove this extraneous output.

This isn't reproducible on a local `make docs` build and may require
further investigation.